Publish to Maven Central
Instead of providing a .aar file that has to be included as a local
dependency, the library is now published to Maven Central.

Publishing is fully automatic and triggered by pushing a tag in the
format x.y.z. Since no official Gradle plugin for publishing to the new
Central Portal exists (as of 2024-03-11), a ZIP distribution is built
and uploaded via cURL using the new Publisher API. Once Sonatype
publishes an official Gradle plugin, this can be changed. See
https://github.com/gradle/gradle/issues/28120 for more details.