Add separate class to hold chip access keys
Currently, there are two ways to authenticate to the chip of the eMRTD:
Using the Card Access Number (CAN) or using details from the MRZ (namely
document number, date of birth, and date of expiry).

Before this commit, the EmrtdConnector#connect(...) method was
overloaded to cover these two ways of authentication. However, this does
not scale very well. In the (very) near future, more options will be
added to the #connect(...) method and having overloads for all possible
combinations of parameters is not feasible.