Add distribution to share with end-user
Before, the user could just checkout the repository from GitHub and run
the test app. Unfortunately, this is no longer possible because they
would need access to our private Maven registry with the emrtd-sdk-java.

Instead there is now a distribution folder with an example app, that
pulls the connector from a public registry. Having a distribution folder
with an example app is the same structure that we use for our offline
SDK.

I don't really like that the test app and the example app are almost
identical and thus share a lot of the same code. I hope I'll find a
better solution in the future. But like I said: It is the same structure
as the offline SDK, so at least it is consistent in that regard...