Autoformat code using Android Studio
Mainly fixing wrong indentation caused by the old .editorconfig file and
by mixing tabs with spaces.

Spaces is the default in Android Studio, why are we using tabs? It's an
invisble character where it is not possible to visually see if a tab or
space is used. Deviating from IDE defaults is just asking for such
mixups to happen.