Improve logging and errors for empty APDU response
At the moment, it is not clear from an error if null, 0 or 1 bytes were
received.

Also according to the IsoDep#transceive(byte[]) JavaDocs, it is not
possible for the method to return null. I didn't want to remove the
check completely (maybe it was there for a reason?), so let's throw an
AssertionError if it happens. It wouldn't be the frist time that
JavaDocs are wrong.