Fix invalid PA on some valid passports
We were using the value of DG14File#getEncoded() to perform Passive
Authentication. This method does reencodes the data group, which
sometimes results in a different byte sequence than the original one
read from the chip.

When checking the hash from the SOD file against this re-encoded byte
array, the PA failed.