Fix and match indent
Replace tabs with spaces, as this is the default setting in Android
Studio and this project contained more files indented with spaces than
with tabs.

Of course, it's a bit silly to use spaces like a four-year-old to create
indentation when there is a character with exactly that semantic meaning.

Nevertheless, other people work with this repository, and frankly, I
don't want to always have to clean up after them.